split write context and write state and add arc for write state
