+++ kernel/src/transaction/bound_write_context.rs
@@ -1,179 +1,21 @@
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::num::NonZero;
 use std::sync::Arc;
 
 use rand::Rng;
-use serde::{Deserialize, Serialize};
 use url::Url;
 
+use super::WriteState;
 use crate::actions::deletion_vector::DeletionVectorPath;
-use crate::expressions::{lit, ColumnName, ExpressionRef, ExpressionStructPatchBuilder, Scalar};
+use crate::expressions::{ColumnName, ExpressionRef};
 use crate::partition::hive::{build_partition_path, uri_encode_path};
-use crate::partition::serialization::serialize_partition_value;
-use crate::partition::validation::validate_partition_values;
-use crate::schema::void_utils::add_void_stripping;
 use crate::schema::SchemaRef;
 use crate::table_features::ColumnMappingMode;
-use crate::utils::require;
-use crate::{DataType, DeltaResult, Error, Expression};
-
-/// Table-wide state required to create [`BoundWriteContext`] instances.
-///
-/// A transaction creates this state once on the driver through
-/// [`Transaction::write_state`](super::Transaction::write_state). Distributed writers can encode
-/// it, transport it to another process, decode it, and bind partition values there without
-/// transporting the transaction itself.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct WriteState {
-    pub(super) table_root: Url,
-    pub(super) full_logical_schema: SchemaRef,
-    pub(super) logical_schema: SchemaRef,
-    pub(super) physical_schema: SchemaRef,
-    pub(super) column_mapping_mode: ColumnMappingMode,
-    pub(super) stats_columns: Vec<ColumnName>,
-    /// Logical partition column names in metadata-defined order.
-    pub(super) logical_partition_columns: Vec<String>,
-    pub(super) materialize_partition_columns: bool,
-    /// Resolved value of the `delta.randomizeFilePrefixes` table property. When true,
-    /// [`BoundWriteContext::write_dir`] emits a random alphanumeric prefix regardless of column
-    /// mapping mode.
-    pub(super) randomize_file_prefixes: bool,
-    /// Resolved value of the `delta.randomPrefixLength` table property. Drives the length
-    /// of the random prefix in [`BoundWriteContext::write_dir`] for both the column mapping and
-    /// `randomizeFilePrefixes` paths.
-    pub(super) random_prefix_length: NonZero<usize>,
-}
-
-impl WriteState {
-    /// Creates a write context for writing data to a specific partition.
-    ///
-    /// The supplied map must contain exactly one entry for every logical partition column. Keys
-    /// are matched case-insensitively and values are validated against the table schema before
-    /// being serialized according to the Delta partition-value rules.
-    ///
-    /// Returns an error if the table is not partitioned or if the partition keys or values are
-    /// invalid.
-    pub fn partitioned_write_context(
-        &self,
-        partition_values: HashMap<String, Scalar>,
-    ) -> DeltaResult<BoundWriteContext> {
-        require!(
-            !self.logical_partition_columns.is_empty(),
-            Error::generic("table is not partitioned; use unpartitioned_write_context() instead")
-        );
-        let normalized = validate_partition_values(
-            &self.logical_partition_columns,
-            &self.full_logical_schema,
-            partition_values,
-        )?;
-
-        let mut serialized = HashMap::with_capacity(normalized.len());
-        for logical_name in &self.logical_partition_columns {
-            let scalar = normalized.get(logical_name).ok_or_else(|| {
-                Error::internal_error(format!(
-                    "partition column '{logical_name}' missing after validation"
-                ))
-            })?;
-            let value = serialize_partition_value(scalar)?;
-            let physical_name = self
-                .full_logical_schema
-                .field(logical_name)
-                .ok_or_else(|| {
-                    Error::internal_error(format!(
-                        "partition column '{logical_name}' not found in schema after validation"
-                    ))
-                })?
-                .physical_name(self.column_mapping_mode)
-                .to_string();
-            serialized.insert(physical_name, value);
-        }
-        let logical_to_physical = Arc::new(self.generate_logical_to_physical(Some(&normalized))?);
-
-        Ok(BoundWriteContext {
-            shared: Arc::new(self.clone()),
-            logical_schema: self.logical_schema.clone(),
-            logical_to_physical,
-            physical_partition_values: serialized,
-        })
-    }
-
-    /// Creates a write context for writing data to an unpartitioned table.
-    ///
-    /// Returns an error if the table has partition columns.
-    pub fn unpartitioned_write_context(&self) -> DeltaResult<BoundWriteContext> {
-        require!(
-            self.logical_partition_columns.is_empty(),
-            Error::generic("table is partitioned; use partitioned_write_context() instead")
-        );
-        let logical_to_physical = Arc::new(self.generate_logical_to_physical(None)?);
-        Ok(BoundWriteContext {
-            shared: Arc::new(self.clone()),
-            logical_schema: self.logical_schema.clone(),
-            logical_to_physical,
-            physical_partition_values: HashMap::new(),
-        })
-    }
-
-    /// Encodes this write state as opaque JSON bytes for transport.
-    ///
-    /// The bytes are tied to this delta-kernel version. Do not inspect them or persist them across
-    /// kernel upgrades.
-    ///
-    /// Returns an error if any field cannot be serialized.
-    pub fn encode(&self) -> DeltaResult<Vec<u8>> {
-        Ok(serde_json::to_vec(self)?)
-    }
-
-    /// Decodes a write state from JSON bytes produced by [`encode`](Self::encode).
-    ///
-    /// The bytes must come from the same delta-kernel version. Cross-version decoding is not
-    /// supported.
-    ///
-    /// Returns an error if the bytes do not contain a valid serialized write state.
-    pub fn decode(bytes: &[u8]) -> DeltaResult<Self> {
-        Ok(serde_json::from_slice(bytes)?)
-    }
-
-    fn generate_logical_to_physical(
-        &self,
-        partition_values: Option<&HashMap<String, Scalar>>,
-    ) -> DeltaResult<Expression> {
-        let mut patch = ExpressionStructPatchBuilder::new();
-        if self.materialize_partition_columns {
-            let partition_cols: HashSet<&str> = self
-                .logical_partition_columns
-                .iter()
-                .map(String::as_str)
-                .collect();
-            let mut predecessor: Option<&str> = None;
-            for field in self.full_logical_schema.fields() {
-                let name = field.name().as_str();
-                if partition_cols.contains(name) {
-                    let value = partition_values
-                        .and_then(|values| values.get(name))
-                        .ok_or_else(|| {
-                            Error::internal_error(format!(
-                                "partition column '{name}' missing while building \
-                                 logical-to-physical expression"
-                            ))
-                        })?;
-                    let literal = lit(value.clone());
-                    patch = match predecessor {
-                        Some(predecessor) => patch.insert_after(predecessor, literal),
-                        None => patch.prepend(literal),
-                    };
-                } else if *field.data_type() != DataType::VOID {
-                    predecessor = Some(name);
-                }
-            }
-        }
-        let patch = add_void_stripping(patch, &self.full_logical_schema);
-        Expression::struct_patch(patch)
-    }
-}
+use crate::{DeltaResult, Error};
 
 /// A write context for a specific partition or an unpartitioned table. Created by either a
-/// [`WriteState`] or the convenience methods on [`Transaction`](super::Transaction).
+/// [`WriteState`](super::WriteState) or the convenience methods on
+/// [`Transaction`](super::Transaction).
 ///
 /// Note: clustered tables are unpartitioned and use `unpartitioned_write_context`.
 ///
@@ -196,8 +38,6 @@ impl WriteState {
 #[derive(Debug)]
 pub struct BoundWriteContext {
     pub(super) shared: Arc<WriteState>,
-    /// Logical schema of the data to write: the table schema minus partition columns.
-    pub(super) logical_schema: SchemaRef,
     /// Transforms logical data to physical data for writing. The logical data must not contain
     /// any partition columns. The expression injects the partition columns when needed.
     pub(super) logical_to_physical: ExpressionRef,
@@ -299,7 +139,7 @@ impl BoundWriteContext {
 
     /// Returns the schema which connectors' logical data should conform to.
     pub fn logical_schema(&self) -> &SchemaRef {
-        &self.logical_schema
+        &self.shared.logical_schema
     }
 
     /// Returns the physical schema (partition columns removed if applicable, column mapping
@@ -435,8 +275,8 @@ mod tests {
     use rstest::rstest;
 
     use super::*;
-    use crate::expressions::{lit, Expression};
-    use crate::schema::{schema_ref, ColumnMetadataKey, MetadataValue, StructField, StructType};
+    use crate::expressions::lit;
+    use crate::schema::schema_ref;
 
     fn make_write_context(
         cm_mode: ColumnMappingMode,
@@ -461,7 +301,6 @@ mod tests {
         });
         BoundWriteContext {
             shared,
-            logical_schema: schema,
             logical_to_physical: Arc::new(lit(true)),
             physical_partition_values: partition_values,
         }
@@ -759,123 +598,5 @@ mod tests {
             Ok(exp) => assert_eq!(wc.resolve_file_path(&file).unwrap(), exp),
             Err(()) => assert!(wc.resolve_file_path(&file).is_err()),
         }
-    }
-
-    fn partitioned_write_state(
-        column_mapping_mode: ColumnMappingMode,
-        materialize_partition_columns: bool,
-        randomize_file_prefixes: bool,
-        random_prefix_length: usize,
-    ) -> WriteState {
-        let year = StructField::not_null("year", DataType::INTEGER).with_metadata([
-            (
-                ColumnMetadataKey::ColumnMappingId.as_ref(),
-                MetadataValue::Number(1),
-            ),
-            (
-                ColumnMetadataKey::ColumnMappingPhysicalName.as_ref(),
-                MetadataValue::String("phys_year".into()),
-            ),
-        ]);
-        let value = StructField::nullable("value", DataType::INTEGER).with_metadata([
-            (
-                ColumnMetadataKey::ColumnMappingId.as_ref(),
-                MetadataValue::Number(2),
-            ),
-            (
-                ColumnMetadataKey::ColumnMappingPhysicalName.as_ref(),
-                MetadataValue::String("phys_value".into()),
-            ),
-        ]);
-        WriteState {
-            table_root: Url::parse("s3://bucket/table/").unwrap(),
-            full_logical_schema: Arc::new(StructType::new_unchecked([year, value])),
-            logical_schema: schema_ref! { nullable "value": INTEGER },
-            physical_schema: schema_ref! { nullable "value": INTEGER },
-            column_mapping_mode,
-            stats_columns: vec![ColumnName::new(["value"])],
-            logical_partition_columns: vec!["year".into()],
-            materialize_partition_columns,
-            randomize_file_prefixes,
-            random_prefix_length: NonZero::new(random_prefix_length).unwrap(),
-        }
-    }
-
-    #[rstest]
-    #[case::default(ColumnMappingMode::None, false, false, 2, "year", false)]
-    #[case::column_mapping(ColumnMappingMode::Name, false, false, 7, "phys_year", true)]
-    #[case::materialized_partition(ColumnMappingMode::None, true, false, 2, "year", false)]
-    #[case::randomized_prefix(ColumnMappingMode::None, false, true, 7, "year", true)]
-    fn write_state_json_round_trip_preserves_worker_behavior(
-        #[case] column_mapping_mode: ColumnMappingMode,
-        #[case] materialize_partition_columns: bool,
-        #[case] randomize_file_prefixes: bool,
-        #[case] random_prefix_length: usize,
-        #[case] expected_partition_key: &str,
-        #[case] expect_random_prefix: bool,
-    ) {
-        let original = partitioned_write_state(
-            column_mapping_mode,
-            materialize_partition_columns,
-            randomize_file_prefixes,
-            random_prefix_length,
-        );
-        let encoded = original.encode().unwrap();
-        let decoded = WriteState::decode(&encoded).unwrap();
-        assert_eq!(decoded.full_logical_schema, original.full_logical_schema);
-        assert_eq!(decoded.logical_schema, original.logical_schema);
-
-        let values = || HashMap::from([("year".to_string(), Scalar::Integer(2024))]);
-        let original_context = original.partitioned_write_context(values()).unwrap();
-        let decoded_context = decoded.partitioned_write_context(values()).unwrap();
-
-        assert_eq!(
-            decoded_context.table_root_dir(),
-            original_context.table_root_dir()
-        );
-        assert_eq!(
-            decoded_context.logical_schema(),
-            original_context.logical_schema()
-        );
-        assert_eq!(
-            decoded_context.physical_schema(),
-            original_context.physical_schema()
-        );
-        assert_eq!(
-            decoded_context.stats_columns(),
-            original_context.stats_columns()
-        );
-        assert_eq!(
-            decoded_context.physical_partition_values(),
-            original_context.physical_partition_values()
-        );
-        assert_eq!(
-            decoded_context.logical_to_physical(),
-            original_context.logical_to_physical()
-        );
-        assert_eq!(decoded_context.column_mapping_mode(), column_mapping_mode);
-        assert_eq!(
-            decoded_context.physical_partition_values(),
-            &HashMap::from([(expected_partition_key.into(), Some("2024".into()))])
-        );
-
-        let write_dir = decoded_context.write_dir().path().to_string();
-        if expect_random_prefix {
-            let prefix = write_dir
-                .strip_prefix("/table/")
-                .unwrap()
-                .strip_suffix('/')
-                .unwrap();
-            assert_eq!(prefix.len(), random_prefix_length);
-            assert!(prefix.chars().all(|c| c.is_ascii_alphanumeric()));
-        } else {
-            assert_eq!(write_dir, "/table/year=2024/");
-        }
-    }
-
-    #[test]
-    fn write_state_decode_rejects_malformed_json() {
-        let error = WriteState::decode(b"not valid json").unwrap_err();
-        assert!(error.to_string().contains("expected ident"));
     }
 }
