+++ kernel/src/transaction/create_table.rs
@@ -176,6 +176,7 @@ impl CreateTableTransaction {
             dv_matched_files: vec![],
             num_dv_updates: 0,
             physical_clustering_columns: clustering_columns,
+            write_state: Default::default(),
             _state: PhantomData,
         })
     }
