+++ kernel/src/transaction/update.rs
@@ -109,6 +109,7 @@ impl Transaction {
             dv_matched_files: vec![],
             num_dv_updates: 0,
             physical_clustering_columns: clustering_columns,
+            write_state: Default::default(),
             _state: PhantomData,
         })
     }
